githubapi: Update for go-github API change.
Follows google/go-github@e48060a28fac52d0f1cb758bc8b87c07bac4a87d.
Updates google/go-github#597.

+++ githubapi/githubapi.go
@@ -51,7 +51,7 @@ func (s service) Get(ctx context.Context, user users.UserSpec) (users.User, erro
 		return users.User{}, fmt.Errorf("user %v not found", user)
 	}
 
-	ghUser, _, err := s.cl.Users.GetByID(ctx, int(user.ID))
+	ghUser, _, err := s.cl.Users.GetByID(ctx, int64(user.ID))
 	if err != nil {
 		return users.User{}, err
 	}
